rename misc_utils -> path_utils. (Better represents function)
