+++ src/SOPRANO/core/analysis.py
@@ -1,7 +1,7 @@
 import pathlib
 
 from SOPRANO.core.objects import AnalysisPaths, GenomePaths
-from SOPRANO.utils.misc_utils import Directories, is_empty
+from SOPRANO.utils.path_utils import Directories, is_empty
 from SOPRANO.utils.sh_utils import pipe
