+++ tests/test_units/test_run_local_ssb_selection.py
@@ -2,16 +2,16 @@ import pathlib
 
 import pytest
 
-import SOPRANO.utils.misc_utils
+import SOPRANO.utils.path_utils
 
 
 def test_check_path():
     # Path is None checks
-    SOPRANO.utils.misc_utils.check_cli_path(None, optional=True)
+    SOPRANO.utils.path_utils.check_cli_path(None, optional=True)
     with pytest.raises(Exception):
-        SOPRANO.utils.misc_utils.check_cli_path(None, optional=False)
+        SOPRANO.utils.path_utils.check_cli_path(None, optional=False)
 
     # Path is not None checks
-    SOPRANO.utils.misc_utils.check_cli_path(pathlib.Path.cwd())
+    SOPRANO.utils.path_utils.check_cli_path(pathlib.Path.cwd())
     with pytest.raises(Exception):
-        SOPRANO.utils.misc_utils.check_cli_path(pathlib.Path.cwd() / "foo")
+        SOPRANO.utils.path_utils.check_cli_path(pathlib.Path.cwd() / "foo")
